Allow desired component/incident status configuration per check.
Some users may want to choose what the incident or component status is
for warning vs critical on a per-check basis, rather than having it
always be `major_outage` and `investigating`. This change provides that
capability, allows for ignoring of certain cases, and ensures only valid
statuses get sent to the statuspage api.